Updating GenerationSizes to be generic.  This allows for maximum generation sizes when generated values larger than Int

+++ src/main/scala/com/github/fulrich/testcharged/generators/numerics/DoubleGenerators.scala
@@ -2,10 +2,10 @@ package com.github.fulrich.testcharged.generators.numerics
 
 
 object DoubleGenerators extends NumericGenerator[Double] {
-  override val Minimum: Int = Double.MinValue.toInt
-  override val TinyMaximum: Int = 25
-  override val ShortMaximum: Int = 1000
-  override val DefaultMaximum: Int = 25000
-  override val BigMaximum: Int = 250000
-  override val HugeMaximum: Int = Double.MaxValue.toInt
+  override val Minimum: Double = Double.MinValue.toInt
+  override val TinyMaximum: Double = 25
+  override val ShortMaximum: Double = 1000
+  override val DefaultMaximum: Double = 25000
+  override val BigMaximum: Double = 250000
+  override val HugeMaximum: Double = Double.MaxValue.toInt
 }
