+++ src/main/scala/com/github/fulrich/testcharged/generators/numerics/FloatGenerators.scala
@@ -2,10 +2,10 @@ package com.github.fulrich.testcharged.generators.numerics
 
 
 object FloatGenerators extends NumericGenerator[Float] {
-  override val Minimum: Int = Float.MinValue.toInt
-  override val TinyMaximum: Int = 25
-  override val ShortMaximum: Int = 1000
-  override val DefaultMaximum: Int = 25000
-  override val BigMaximum: Int = 250000
-  override val HugeMaximum: Int = Float.MaxValue.toInt
+  override val Minimum: Float = Float.MinValue.toInt
+  override val TinyMaximum: Float = 25
+  override val ShortMaximum: Float = 1000
+  override val DefaultMaximum: Float = 25000
+  override val BigMaximum: Float = 250000
+  override val HugeMaximum: Float = Float.MaxValue.toInt
 }
