+++ src/main/scala/com/github/fulrich/testcharged/generators/numerics/ShortGenerators.scala
@@ -2,10 +2,10 @@ package com.github.fulrich.testcharged.generators.numerics
 
 
 object ShortGenerators extends NumericGenerator[Short] {
-  override val Minimum: Int = Short.MinValue
-  override val TinyMaximum: Int = 25
-  override val ShortMaximum: Int = 250
-  override val DefaultMaximum: Int = 5000
-  override val BigMaximum: Int = 20000
-  override val HugeMaximum: Int = Short.MaxValue
+  override val Minimum: Short = Short.MinValue
+  override val TinyMaximum: Short = 25
+  override val ShortMaximum: Short = 250
+  override val DefaultMaximum: Short = 5000
+  override val BigMaximum: Short = 20000
+  override val HugeMaximum: Short = Short.MaxValue
 }
